+++ backend/packages/harness/deerflow/persistence/models/__init__.py
@@ -1,6 +1,21 @@
-from deerflow.persistence.models.feedback import FeedbackRow
-from deerflow.persistence.models.run import RunRow
+"""ORM model registration entry point.
+
+Importing this module ensures all ORM models are registered with
+``Base.metadata`` so Alembic autogenerate detects every table.
+
+The actual ORM classes have moved to entity-specific subpackages:
+- ``deerflow.persistence.thread_meta``
+- ``deerflow.persistence.run``
+- ``deerflow.persistence.feedback``
+
+``RunEventRow`` remains in ``deerflow.persistence.models.run_event`` because
+its storage implementation lives in ``deerflow.runtime.events.store.db`` and
+there is no matching entity directory.
+"""
+
+from deerflow.persistence.feedback.model import FeedbackRow
 from deerflow.persistence.models.run_event import RunEventRow
-from deerflow.persistence.models.thread_meta import ThreadMetaRow
+from deerflow.persistence.run.model import RunRow
+from deerflow.persistence.thread_meta.model import ThreadMetaRow
 
 __all__ = ["FeedbackRow", "RunEventRow", "RunRow", "ThreadMetaRow"]
