--- backend/packages/harness/deerflow/persistence/repositories/__init__.py
@@ -1,5 +1,0 @@
-from deerflow.persistence.repositories.feedback_repo import FeedbackRepository
-from deerflow.persistence.repositories.run_repo import RunRepository
-from deerflow.persistence.repositories.thread_meta_repo import ThreadMetaRepository
-
-__all__ = ["FeedbackRepository", "RunRepository", "ThreadMetaRepository"]
